src/main/java/com/es/phoneshop/dao/impl/OrderDaoImpl.java
package com.es.phoneshop.dao.impl;

import com.es.phoneshop.dao.OrderDao;
import com.es.phoneshop.model.order.Order;

import java.util.ArrayList;
import java.util.List;

public class OrderDaoImpl extends GenericDaoImpl<Order> implements OrderDao {

    private static OrderDaoImpl instance;

    private List<Order> orders;

    private OrderDaoImpl() {
        this.orders = new ArrayList<>();
    }

    public static synchronized OrderDaoImpl getInstance() {
        if (instance == null) {
            instance = new OrderDaoImpl();
        }
        return instance;
    }

    public List<Order> getOrders() {
        return orders;
    }

    public void setOrders(List<Order> orders) {
        this.orders = orders;
    }

    @Override
    public long getId(Order order) {
        return order.getId();
    }
}
